requirements.txt
SQLAlchemy==1.0.12
aiohttp==0.19.0
beautifulsoup4==4.3.2
bs4==0.0.1
cassiopeia==0.1.1
chardet==2.3.0
colorama==0.3.3
command-not-found==0.3
discord.py==0.10.0a0
html5lib==0.999
language-selector==0.1
lxml==3.4.4
pycurl==7.19.5.1
pygobject==3.16.2
python-apt===1.0.1-ubuntu0.1
python-debian==0.1.27
requests==2.7.0
six==1.9.0
ssh-import-id==4.5
ufw==0.34
unattended-upgrades==0.1
urllib3==1.11
websockets==2.7
wheel==0.26.0
youtube-dl==2016.4.19
